chore: move usePeople related to framework hosted
