+++ src/social-network-provider/facebook.com/ui-provider.ts
@@ -1,9 +1,14 @@
 import { defineSocialNetworkUI } from '../../social-network/ui'
+import { ValueRef } from '@holoflows/kit/es'
+import { InitFriendsValueRef } from '../../social-network/defaults/FriendsValueRef'
 
 defineSocialNetworkUI({
     name: 'Facebook',
-    init(env, pref) {},
+    init(env, pref) {
+        InitFriendsValueRef(this, 'facebook.com')
+    },
     shouldActivate() {
         return location.hostname.endsWith('facebook.com')
     },
+    friendsRef: new ValueRef([]),
 })
